Renamed relation package to predicates
