+++ src/main/java/org/mutation_testing/predicates/RelationType.java
@@ -1,4 +1,4 @@
-package org.mutation_testing.relation;
+package org.mutation_testing.predicates;
 
 public enum RelationType {
     NAME_LITERAL, NAME_NAME, NAME_EXPR, EXPR_EXPR, EXPR_LITERAL, LITERAL_LITERAL;
